wr_malloc_size_of

Trait MallocShallowSizeOf

pub trait MallocShallowSizeOf {
    // Required method
    fn shallow_size_of(&self, ops: &mut MallocSizeOfOps) -> usize;
}

Trait for measuring the “shallow” heap usage of a container.

Required Methods

fn shallow_size_of(&self, ops: &mut MallocSizeOfOps) -> usize

Measure the heap usage of immediate heap-allocated descendant structures, but not the space taken up by the value itself. Anything beyond the immediate descendants must be measured separately, using iteration.