Crate wr_malloc_size_of

A reduced fork of Firefox's malloc_size_of crate, for bundling with WebRender.

Macros

malloc_size_of_is_0
For use on types where size_of() returns 0.

Structs

MallocSizeOfOps
Operations used when measuring heap usage of data structures.

Traits

MallocShallowSizeOf
Trait for measuring the "shallow" heap usage of a container.

MallocSizeOf
Trait for measuring the "deep" heap usage of a data structure. This is the most commonly-used of the traits.

Type Definitions

VoidPtrToSizeFn
A C function that takes a pointer to a heap allocation and returns its size.